Refs #31349. Update tests.
